# docs/swagger.yaml
definitions:
...
  fitness-tracker-backend_workout_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.trainingListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Training'
        type: array
      next_cursor:
        type: string
    type: object
  handler.trainingRequest:
    properties:
      finished_at:
        type: string
      notes:
        type: string
      sessions:
        items:
          $ref: '#/definitions/handler.trainingSessionRequest'
        type: array
      started_at:
        type: string
      title:
        type: string
    type: object
  handler.trainingSessionRequest:
    properties:
      datetime:
        type: string
      order_index:
        type: integer
      sets:
        items:
          $ref: '#/definitions/handler.workoutSetRequest'
        type: array
      workout_type_id:
        type: integer
    required:
    - workout_type_id
    type: object
  handler.trainingUpdateRequest:
    properties:
      finished_at:
        type: string
      notes:
        type: string
      started_at:
        type: string
      title:
        type: string
    type: object
...
  handler.workoutSetRequest:
    properties:
      distance_meters:
        type: number
      duration_seconds:
        type: integer
      notes:
        type: string
      reps:
        type: integer
      set_number:
        type: integer
      weight_kg:
        type: number
    type: object
...
  models.MuscleGroup:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Training:
    properties:
      created_at:
        type: string
      finished_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      sessions:
        items:
          $ref: '#/definitions/models.WorkoutSession'
        type: array
      started_at:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.WorkoutSession:
    properties:
      datetime:
        type: string
      id:
        type: integer
      order_index:
        type: integer
      sets:
        items:
          $ref: '#/definitions/models.WorkoutSet'
        type: array
      training:
        allOf:
        - $ref: '#/definitions/models.Training'
        description: Associations
      training_id:
        type: integer
      user_id:
        type: integer
      workout_type:
        $ref: '#/definitions/models.WorkoutType'
      workout_type_id:
        type: integer
    type: object
  models.WorkoutSet:
    properties:
      distance_meters:
        type: number
      duration_seconds:
        type: integer
      id:
        type: integer
      notes:
        type: string
      reps:
        type: integer
      set_number:
        type: integer
      weight_kg:
        type: number
      workout_session_id:
        type: integer
    type: object
  models.WorkoutType:
    properties:
      default_metric:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      muscle_group:
        allOf:
        - $ref: '#/definitions/models.MuscleGroup'
        description: Associations
      muscle_group_id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for the Fitness Tracker backend service.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /trainings:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: RFC3339 started_at cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.trainingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List trainings for current user
      tags:
      - trainings
    post:
      consumes:
      - application/json
      parameters:
      - description: Training
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.trainingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create training with nested sessions and sets
      tags:
      - trainings
  /trainings/{id}:
    delete:
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete training
      tags:
      - trainings
    get:
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get training by ID
      tags:
      - trainings
    patch:
      consumes:
      - application/json
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.trainingUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Training'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fitness-tracker-backend_workout_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update training metadata
      tags:
      - trainings
...
securityDefinitions:
  BearerAuth:
    description: 'Provide your JWT with the "Bearer " prefix. Example: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"